first commit but it run well

+++ code/sink/attributed_checkouts.sql
@@ -5,7 +5,7 @@ CREATE TABLE attributed_checkouts (
     click_id STRING,
     product_id STRING,
     payment_method STRING,
-    total_amount DECIMAL(5, 2),
+    total_amount FLOAT,
     shipping_address STRING,
     billing_address STRING,
     user_agent STRING,
